+++ CHANGELOG.md
@@ -13,6 +13,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-11
+### Added
+- Warnings for missing method arguments in docstrings.
+- Docstring comparisons.
+
+## [2.0.2] - 2023-10
+### Added
+- Support for Python 3.12.
+ 
 ## [2.0.1] - 2023-09
 ### Fixed
 - Parsing of Google docstrings with Sphinx directives (like `.. math::`) is no longer considered as a section.